# .mcp/spring-ai-openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:18080","description":"Generated server url"}],"paths":{"/api/tts":{"post":{"tags":["text-to-speech-controller"],"operationId":"synthesize","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextToSpeechRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/rag/query":{"post":{"tags":["rag-controller"],"operationId":"query","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RagQueryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RagResponse"}}}}}}},"/api/rag/documents":{"get":{"tags":["rag-controller"],"operationId":"listDocuments","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSummary"}}}}}}},"post":{"tags":["rag-controller"],"operationId":"addDocuments","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentPayload"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSummary"}}}}}}}},"/api/chat":{"post":{"tags":["chat-controller"],"operationId":"prompt","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResult"}}}}}}},"/api/chat/ocr":{"post":{"tags":["chat-controller"],"operationId":"ocr","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"payload":{"$ref":"#/components/schemas/OcrRequest"},"media":{"type":"string","format":"binary"}},"required":["media","payload"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OcrResult"}}}}}}},"/api/chat/multimodal":{"post":{"tags":["chat-controller"],"operationId":"promptWithMedia","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"payload":{"$ref":"#/components/schemas/MultimodalRequest"},"media":{"type":"string","format":"binary"}},"required":["payload"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResult"}}}}}}},"/api/chat/memory/{conversationId}":{"get":{"tags":["chat-controller"],"operationId":"memory","parameters":[{"name":"conversationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatMemorySnapshot"}}}}}}}},"components":{"schemas":{"TextToSpeechRequest":{"type":"object","properties":{"text":{"type":"string","minLength":1},"voice":{"type":["string","null"]},"format":{"type":["string","null"]},"speed":{"type":["number","null"],"format":"double","maximum":2.0,"minimum":0.5}},"required":["text"]},"RagQueryRequest":{"type":"object","properties":{"question":{"type":"string","minLength":1},"topK":{"type":["integer","null"],"format":"int32","maximum":50,"minimum":1},"similarityThreshold":{"type":["number","null"],"format":"double","maximum":1.0,"minimum":0.0},"filterExpression":{"type":["string","null"]}},"required":["question"]},"DocumentSummary":{"type":"object","properties":{"id":{"type":"string"},"content":{"type":"string"},"metadata":{"type":"object","additionalProperties":{}}},"required":["content","id","metadata"]},"RagResponse":{"type":"object","properties":{"answer":{"type":"string"},"context":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSummary"}}},"required":["answer","context"]},"DocumentPayload":{"type":"object","properties":{"id":{"type":["string","null"]},"text":{"type":"string","minLength":1},"source":{"type":["string","null"]},"metadata":{"type":["object","null"],"additionalProperties":{"type":"string"}}},"required":["text"]},"ChatRequest":{"type":"object","properties":{"message":{"type":"string","minLength":1},"voice":{"type":["string","null"]},"conversationId":{"type":["string","null"]}},"required":["message"]},"ChatResult":{"type":"object","properties":{"message":{"type":"string"},"voice":{"type":"string"}},"required":["message","voice"]},"OcrRequest":{"type":"object","properties":{"instructions":{"type":"string","minLength":1},"voice":{"type":["string","null"]},"conversationId":{"type":["string","null"]}},"required":["instructions"]},"OcrResult":{"type":"object","properties":{"extractedText":{"type":"string"},"voice":{"type":"string"}},"required":["extractedText","voice"]},"MultimodalRequest":{"type":"object","properties":{"message":{"type":"string","minLength":1},"voice":{"type":["string","null"]},"conversationId":{"type":["string","null"]}},"required":["message"]},"ChatMemorySnapshot":{"type":"object","properties":{"conversationId":{"type":"string"},"messages":{"type":"array","items":{"type":"string"}}},"required":["conversationId","messages"]}}}}